fix private key length issues by always doing setLengthLeft
Summary:
When representing a private key, signature or operation in hex, we should always do EthJSUtil.setLengthLeft(...).
If not, many parsers will consider the length of the resultant private key or signature to be invalid!
This bug is common across many ethereum libraries (i encountered it in testrpc myself, and it is a live bug in lightwallet).

A good time to consider this is whenever you are doing low level operations to do with keys / sigs and are using toString('hex').

Reviewers: ben, huang, alex, arik

Subscribers: ben

Differential Revision: https://phabricator.bitgo.com/D3121